prebias: carry the logical position as an integer
The position base was a float tensor because that reused plumbing
that already worked, guarded by a check that rejected anything at or
beyond two to the twenty-fourth.  That put a floating-point exactness
boundary inside a contract about sequence positions.

Declare the tensor as int32 and require an integer dtype, accepting
int64 and narrowing it.  Positions past the old limit now work, and
a float tensor is refused rather than silently truncated.

+++ flashinfer/jit/attention/modules.py
@@ -1062,10 +1062,11 @@ def gen_batch_prefill_module(
             additional_tensor_dtypes.append("float")
             additional_tensor_names.append("maybe_prebias_rope_table")
             additional_tensor_dtypes.append("float")
-            # per-request logical start position; float keeps the existing
-            # plumbing and is exact for any position below 2^24
+            # per-request logical start position. A position is an integer;
+            # carrying it as float worked only below 2^24 and made the
+            # contract depend on a floating-point exactness boundary.
             additional_tensor_names.append("maybe_prebias_pos_base")
-            additional_tensor_dtypes.append("float")
+            additional_tensor_dtypes.append("int32_t")
         additional_scalar_names = [
             "logits_soft_cap",
             "sm_scale",
